+++ NEWS.md
@@ -1,2 +1,4 @@
 0.1.0 
-* Initial release
+=========
+
+- Using the `cloudyr` `aws.signature` package vs DIY